requirements.txt
streamlit
numpy
pandas
matplotlib
Pillow
scikit-learn
